python-data-engineer/week8/mini_project/requirements.txt
# Core
pyyaml==6.0.1
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# Streaming
kafka-python==2.0.2
apache-flink==1.17.1
debezium-connector-python==1.0.0

# API Clients
shopify==12.3.0
woocommerce==3.0.0

# Processing
pandas==2.1.3
numpy==1.26.2

# Monitoring
prometheus-client==0.19.0
grafana-api-client==2.1.0
alertmanager-client==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0 